+++ elmyra.ip.access.epo/elmyra/ip/access/epo/helpers.py
@@ -2,12 +2,15 @@
 # (c) 2014 Andreas Motl, Elmyra UG
 import json
 import logging
-from pyramid.settings import asbool
+from pyramid.settings import asbool     # required by template
 from pyramid.threadlocal import get_current_request
 
 log = logging.getLogger(__name__)
 
 class BackboneModelParameterFiddler(object):
+    """all parameter fiddling in one single place :-)"""
+
+    # TODO: refactor IpsuiteNavigatorConfig.defaults here as well, trim down config.js
 
     def __init__(self, name):
         self.name = name
@@ -34,7 +37,6 @@ class BackboneModelParameterFiddler(object):
         return realdata
 
     def compute_parameters(self):
-        """all parameter lifting in one place"""
 
         request = get_current_request()
 
@@ -42,14 +44,18 @@ class BackboneModelParameterFiddler(object):
         request_params = dict(request.params)
         request_opaque = dict(request.opaque)
 
-        # parameter firewall
+
+        # A. parameter firewall
+        host = request.headers.get('Host')
+        isviewer = host == 'patentview.elmyra.de'
+
         # 1. don't allow "query" from outside on viewer-only domains
-        query_allowed = request.headers.get('Host') != 'patentview.elmyra.de'
-        if request_params.has_key('query') and not query_allowed:
+        if request_params.has_key('query') and isviewer:
             log.warn('parameter "query=%s" not allowed on this instance, deleting it', request_params['query'])
             del request_params['query']
 
-        # merge parameters
+
+        # B. merge parameters
         # 1. use "settings" as foundation
         # 2. merge "request parameters"
         # 3. merge "opaque parameters" taking the highest precedence
@@ -57,7 +63,16 @@ class BackboneModelParameterFiddler(object):
         params.update(request_params)
         params.update(request_opaque)
 
-        # backward-compat amendments
+
+        # C. special customizations
+
+        # 1. on patentview.elmyra.de, only run liveview
+        if isviewer:
+            params['mode'] = 'liveview'
+            params['setting.ui.page.title'] = 'Patent view'
+
+
+        # D. backward-compat amendments
         for key, value in params.iteritems():
             if key.startswith('ship_'):
                 newkey = key.replace('ship_', 'ship-')