+++ elmyra.ip.access.epo/CHANGES.rst
@@ -5,6 +5,7 @@ elmyra.ip.access.epo CHANGES
 development
 ===========
 - middleware: attempt to fix url routing re. undesired infinite redirect loops
+- middleware: bind host=patentview.elmyra.de to mode=liveview and title=Patent view
 
 0.21.1
 ======